use int

+++ test/element_access.cpp
@@ -148,7 +148,7 @@ BOOST_AUTO_TEST_CASE(empty_elements) {
 }
 
 BOOST_AUTO_TEST_CASE(multi_test_elements_1D) {
-	multi::array<double, 1> arr = {0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0};
+	multi::array<int, 1> arr = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
 	BOOST_REQUIRE( arr.size() == 10 );
 
 	BOOST_REQUIRE(  arr.elements().size() == 10 );
@@ -169,32 +169,32 @@ BOOST_AUTO_TEST_CASE(multi_test_elements_1D) {
 	BOOST_REQUIRE(  arr().elements().end()  >  arr().elements().begin() );
 	BOOST_REQUIRE(  arr().elements().end()  >= arr().elements().begin() );
 
-	arr.elements() = {9.0, 8.0, 7.0, 6.0, 5.0, 4.0, 3.0, 2.0, 1.0, 0.0};
-	BOOST_REQUIRE( arr[2] == 7.0 );
-	BOOST_REQUIRE( arr.elements()[2] == 7.0 );
+	arr.elements() = {9, 8, 7, 6, 5, 4, 3, 2, 1, 0};
+	BOOST_REQUIRE( arr[2] == 7 );
+	BOOST_REQUIRE( arr.elements()[2] == 7 );
 }
 
 BOOST_AUTO_TEST_CASE(multi_test_elements_1D_as_range) {
-	multi::array<double, 1> arr = {0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0};
+	multi::array<int, 1> arr = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
 	BOOST_REQUIRE( arr.size() == 10 );
 
-	arr().elements() = {9.0, 8.0, 7.0, 6.0, 5.0, 4.0, 3.0, 2.0, 1.0, 0.0};
-	BOOST_REQUIRE( arr[2] == 7. );
-	BOOST_REQUIRE( arr.elements()[2] == 7.0 );
+	arr().elements() = {9, 8, 7, 6, 5, 4, 3, 2, 1, 0};
+	BOOST_REQUIRE( arr[2] == 7 );
+	BOOST_REQUIRE( arr.elements()[2] == 7 );
 }
 
 BOOST_AUTO_TEST_CASE(elements_from_init_list_2D) {
-	multi::array<double, 2> arr({3, 2});
-	arr().elements() = {1.0, 2.0, 3.0, 4.0, 5.0, 6.0};
-	BOOST_REQUIRE(arr[1][0] == 3.0);
+	multi::array<int, 2> arr({3, 2});
+	arr().elements() = {1, 2, 3, 4, 5, 6};
+	BOOST_REQUIRE(arr[1][0] == 3);
 
-	arr.elements() = {10.0, 20.0, 30.0, 40.0, 50.0, 60.0};
+	arr.elements() = {10, 20, 30, 40, 50, 60};
 	BOOST_REQUIRE(arr[1][0] == 30.);
 }
 
 BOOST_AUTO_TEST_CASE(front_back_2D) {
-	multi::array<double, 2> arr({3, 4});
-	std::iota(arr.data_elements(), arr.data_elements() + arr.num_elements(), 0.0);
+	multi::array<int, 2> arr({3, 4});
+	std::iota(arr.data_elements(), arr.data_elements() + arr.num_elements(), 0);
 
 	BOOST_REQUIRE(  arr.front()[2] ==  arr[0][2] );
 	BOOST_REQUIRE( &arr.front()[2] == &arr[0][2] );
@@ -204,8 +204,8 @@ BOOST_AUTO_TEST_CASE(front_back_2D) {
 }
 
 BOOST_AUTO_TEST_CASE(front_back_1D) {
-	multi::array<double, 1> arr({30}, double{});
-	std::iota(arr.data_elements(), arr.data_elements() + arr.num_elements(), 0.0);
+	multi::array<int, 1> arr({30}, double{});
+	std::iota(arr.data_elements(), arr.data_elements() + arr.num_elements(), 0);
 
 	BOOST_REQUIRE(  arr.front() ==  arr[ 0] );
 	BOOST_REQUIRE( &arr.front() == &arr[ 0] );
@@ -215,8 +215,8 @@ BOOST_AUTO_TEST_CASE(front_back_1D) {
 }
 
 BOOST_AUTO_TEST_CASE(elements_rvalues) {
-	using movable_type = std::vector<double>;
-	movable_type const movable_value(5., 99.0);  // NOLINT(fuchsia-default-arguments-calls)
+	using movable_type = std::vector<int>;
+	movable_type const movable_value(5, 99);  // NOLINT(fuchsia-default-arguments-calls)
 
 	multi::array<movable_type, 1> arr = {movable_value, movable_value, movable_value};
 	BOOST_REQUIRE( arr.size() == 3 );
